+++ CHANGELOG.md
@@ -11,15 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -
 -
 -
--
+- Add `plot_change_points_interactive` ([#988](https://github.com/tinkoff-ai/etna/pull/988))
 -
 -
 -
 -
 ### Changed
 -
--
+- Change returned model in get_model of BATSModel, TBATSModel ([#987](https://github.com/tinkoff-ai/etna/pull/987))
 - 
+-
+- Change returned model in `get_model` of `HoltWintersModel`, `HoltModel`, `SimpleExpSmoothingModel` ([#986](https://github.com/tinkoff-ai/etna/pull/986))
 - 
 - 
 -